map_flywheel_to_redcap_gear/requirements.txt
flywheel-gear-toolkit
flywheel-sdk==12.4.0
jsonschema==3.2.0
numpy==1.19.2
pandas==1.1.2
pathlib==1.0.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3
zipp==3.1.0
pyyaml
pycap